mcp-agent/requirements.txt
# MCP Server Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
